requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
python-telegram-bot==20.0
requests==2.28.1
python-dotenv==0.21.0
werkzeug==2.2.2
gunicorn==20.1.0